Update to dev
The devel branch has just a few commits squash some bugs that haven't made it to master and since it's been a year with no updates I'm not sure it will be anytime soon. As a side note maybe look to see if someone else has picked up coding on this.